# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "manjunathhalli629@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/forgotpassword": {
            "post": {
                "summary": "forgot password",
                "description": "forgot user password",
                "operationId": "186a73907a51e79b61ee87564453da94",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset password link sent to email id"
                    },
                    "404": {
                        "description": "wrong user with thiss email id"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/resetpassword": {
            "post": {
                "summary": "reset password",
                "description": "reset user password",
                "operationId": "70a80c24319db6e6fd175bbab31d79bc",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "new_password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Password reset successfull!"
                    },
                    "400": {
                        "description": "wrong user with this Email ID"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}